requirements.txt
# Core dependencies for Stable Flow
Jinja2>=3.1.0
PyYAML>=6.0
Click>=8.0.0
colorama>=0.4.4

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional dependencies for advanced features
rich>=12.0.0
typer>=0.7.0
